Template content tests
Changed some 'pending' template tests to check for attributes, defaults,
and platform specific changes.
Also, noticed that the spec memcached servers list was a string,
and the template uses .each, so I changed it to an array.
And changed the Berksfile to use stackforge

Also, updated spec tests to use new cookbook name 'openstack-dashboard'
and found 1 lingering 'horizon' attribute in dash-site template

Change-Id: I80d57c6f37583aec374f837c9b832576ac0e8b1b