Do not configure memcache if no servers found
If a user specified memcache as the session_backend for the dashboard
then make sure we have memcache servers to configure local_settings.py
with. Attempting to configure this Django front-end with memcache on
and no servers found/specified will cause Django to 500. But, that is
not to say that this configuration is an error - this could happen as
the environment is being converged across nodes.

This change adds tests to support this logic.

Change-Id: I72cf5465d09aec34d9a76305819745b0580b19a2