creates openstack-dashboard site properly on centos/rhel
Fixes: bug #1271193
Implements: blueprint centos-rhel-for-dashboard

On centos/rhel we now support the sites-enabled/sites-available style
confs.  This patch moves to use this approach on centos/rhel. It also
fixes the above bug which was causing the site not to get loaded
properly because it did not have the .conf extension. Also remove the
rhel/centos specific tests for the openstack-dashboard site since
rhel/centos now behaves the same as ubuntu in this regard.

Note I am purposely ignoring fedora/others in this patch since the level
of support for those is still tbd.

Change-Id: I7c3339b7e3a5cd019b78dbb05425fef6fb344ebc