Only run the syncdb when set the session backend to sql.
From the horizon manual
http://docs.openstack.org/developer/horizon/topics/deployment.html#database,
we need the database only when set the session backend to sql, so we
just need run syncdb at that case.
Another reason I don't want to run this syncdb is, normally the web
application (here is the horizon) should not access the database server
directly in data center for security consideration.

Change-Id: I39323fc797bcb9a9b821e955d8f648a4271086b1