Fix site directory defaults for apache 2.4
With the 2.4 release the site config for directies was changed as
noted in the bug link.  If necessary, in the future we could create
a list attribute to hold the directory options to allow them to be
overrideable, but no one has asked for that yet.  This could also
use the apache cookbook web_app provider to create this site config.
Also cleaned up some minor formatting in the specs.

Change-Id: Ie2b323ac055ae03b3c4ebb28b70aca65745df842
Closes-Bug: #1411772