Include ServerAlias in dashboard vhost if set
There are use cases where having ServerAlias in the vhost is preferred. This
includes that option but excludes by default. It's set using an array so you can
set multiple server aliases.

I also added the virtualhost port configurator test for dashboard-http-bind on
port 80 to ensure it works on both ports.

Change-Id: I621cb2a519e92d04d60fd75e727b59250bd72e30